Use separate namespaces for offset bit widths
Namespaces pid8, pid16, pid32, pid64 contain structs which use signed
integers with 8, 16, 32, 64 bits as the offset, respectively.

Some examples:

pid32::ptr<T> is a relative pointer with a 32-bit signed integer offset,
and similarly for the other bit widths.

pid32::string16 is a string, which uses a 32-bit offset to its data, and
the maximum size of the data is 65535 bytes (2^16 - 1).

pid64::vector32 is a vector, which uses a 64-bit offset to its data, and
the maximum number of items is 2^32 - 1.

pid16::map8<pid32::string32, std::int32_t> is a map, which uses a 16-bit
offset to its data. The keys are strings which use 32-bit offsets to
their data, and which are at most 2^32 - 1 bytes long. The values are
32-bit signed integers.